+++ app/src/main/java/org/simple/clinic/bp/entry/BloodPressureEntrySheet.kt
@@ -40,6 +40,7 @@ import org.simple.clinic.widgets.UiEvent
 import org.simple.clinic.widgets.ageanddateofbirth.UserInputDateValidator
 import org.simple.clinic.widgets.displayedChildResId
 import org.simple.clinic.widgets.setTextAndCursor
+import org.simple.clinic.widgets.visibleOrGone
 import java.time.LocalDate
 import java.time.ZoneId
 import java.time.format.DateTimeFormatter
@@ -387,6 +388,22 @@ class BloodPressureEntrySheet : BottomSheetActivity(), BloodPressureEntryUi, Rem
 
   override fun dismiss() {
     finish()
+  }
+
+  override fun showProgress() {
+    progressLoader.visibleOrGone(isVisible = true)
+    bloodPressureEntrySpace.visibleOrGone(isVisible = true)
+    bloodPressureEntryLayout.visibleOrGone(isVisible = false)
+    bpDateButton.visibleOrGone(isVisible = false)
+    removeBloodPressureButton.visibleOrGone(isVisible = false)
+  }
+
+  override fun hideProgress() {
+    progressLoader.visibleOrGone(isVisible = false)
+    bloodPressureEntrySpace.visibleOrGone(isVisible = false)
+    bloodPressureEntryLayout.visibleOrGone(isVisible = true)
+    bpDateButton.visibleOrGone(isVisible = true)
+    removeBloodPressureButton.visibleOrGone(isVisible = true)
   }
 
   override fun onBloodPressureRemoved() {