+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add progress UI for registering a patient
 - Add progress UI when scheduling appointment in `ScheduleAppointmentSheet`
 - Add progress UI when editing a patient in `EditPatientScreen`
+- Add progress UI when saving Blood pressure measurements in `BloodPressureEntrySheet`
 
 ### Changes
 - Updated translations: `ti`